Admin.TagOverview.feature backend lint fixes
